관리자 로그인 밖에 없으니까, 세션에 의한 로그인 시스템 채택

+++ src/main/java/project/MilkyWay/Address/Controller/AddressController.java
@@ -5,6 +5,7 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -14,6 +15,9 @@ import project.MilkyWay.ComonType.DTO.ResponseDTO;
 import project.MilkyWay.Address.Entity.AddressEntity;
 import project.MilkyWay.ComonType.Expection.FindFailedException;
 import project.MilkyWay.Address.Service.AddressService;
+import project.MilkyWay.ComonType.Expection.SessionNotFoundExpection;
+import project.MilkyWay.ComonType.Expection.UpdateFailedException;
+import project.MilkyWay.ComonType.LoginSuccess;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -30,30 +34,41 @@ public class AddressController
 
     ResponseDTO<AddressDTO> responseDTO = new ResponseDTO<>();
 
+    LoginSuccess loginSuccess = new LoginSuccess();
+
+
     @Operation(
             summary = "Create a new Address",
-            description = "This API creates a new Address and returns AddressDTO as response",
+            description = "This API creates a new Address and returns AddressDTO as response, but only if the user is an administrator.",
             responses = {
                     @ApiResponse(responseCode = "201", description = "Address created successfully", content = @Content(mediaType = "application/json", schema = @Schema(implementation = AddressDTO.class))),
                     @ApiResponse(responseCode = "400", description = "Invalid input data")
             }
     )
     @PostMapping("/Insert")
-    public ResponseEntity<?> Insert(@Valid @RequestBody AddressDTO addressDTO)
+    public ResponseEntity<?> Insert(HttpServletRequest request, @Valid @RequestBody AddressDTO addressDTO)
     {
         try
         {
-            AddressEntity addressEntity = ConvertToEntity(addressDTO);
-            AddressEntity addressEntity1 = addressService.insert(addressEntity);
-            if(addressEntity1 != null)
+            if(loginSuccess.isSessionExist(request))
             {
-                AddressDTO addressDTO1 = ConvertToDTO(addressEntity1);
-                return ResponseEntity.ok().body(responseDTO.Response("success","데이터베이스에 주소 데이터 추가", Collections.singletonList(addressDTO1)));
+                AddressEntity addressEntity = ConvertToEntity(addressDTO);
+                AddressEntity addressEntity1 = addressService.insert(addressEntity);
+                if(addressEntity1 != null)
+                {
+                    AddressDTO addressDTO1 = ConvertToDTO(addressEntity1);
+                    return ResponseEntity.ok().body(responseDTO.Response("success","데이터베이스에 주소 데이터 추가", Collections.singletonList(addressDTO1)));
+                }
+                else
+                {
+                    throw new RuntimeException("예기치 못한 오류로 런타임 오류 발생!!");
+                }
             }
             else
             {
-                throw new RuntimeException("예기치 못한 오류로 런타임 오류 발생!!");
+                throw new SessionNotFoundExpection("관리자 로그인 X, 주소 정보를 추가할 수 없습니다.");
             }
+
         }
         catch (Exception e)
         {
@@ -63,7 +78,7 @@ public class AddressController
 
     @Operation(
             summary = "Change an Address by AddressId",
-            description = "This API changes an Address and returns AddressDTO as response",
+            description = "This API changes an Address and returns AddressDTO as response, but only if the user is an administrator.",
             responses = {
                     @
                     ApiResponse(responseCode = "201", description = "Address Changed successfully", content = @Content(mediaType = "application/json", schema = @Schema(implementation = AddressDTO.class))),
@@ -71,21 +86,29 @@ public class AddressController
             }
     )
     @PutMapping("/Update")
-    public ResponseEntity<?> Update(@Valid @RequestBody AddressDTO addressDTO)
+    public ResponseEntity<?> Update(HttpServletRequest request, @Valid @RequestBody AddressDTO addressDTO)
     {
         try
         {
-            AddressEntity addressEntity = ConvertToEntity(addressDTO);
-            AddressEntity addressEntity1 = addressService.update(addressEntity);
-            if(addressEntity1 != null)
+            if(loginSuccess.isSessionExist(request))
             {
-                AddressDTO addressDTO1 = ConvertToDTO(addressEntity1);
-                return ResponseEntity.ok().body(responseDTO.Response("success","데이터베이스에 주소 데이터 수정", Collections.singletonList(addressDTO1)));
+                AddressEntity addressEntity = ConvertToEntity(addressDTO);
+                AddressEntity addressEntity1 = addressService.update(addressEntity);
+                if(addressEntity1 != null)
+                {
+                    AddressDTO addressDTO1 = ConvertToDTO(addressEntity1);
+                    return ResponseEntity.ok().body(responseDTO.Response("success","데이터베이스에 주소 데이터 수정", Collections.singletonList(addressDTO1)));
+                }
+                else
+                {
+                    throw new UpdateFailedException("청소 예약 정보 등록 과정에서 오류 발생!! 관리자에게 문의하세요");
+                }
             }
             else
             {
-                throw new RuntimeException("예기치 못한 오류로 런타임 오류 발생!!");
+                throw new SessionNotFoundExpection("관리자 로그인X, 다시 로그인을 시도하여 주세요");
             }
+
         }
         catch (Exception e)
         {
@@ -95,7 +118,7 @@ public class AddressController
     }
 
     @Operation(
-            summary = "Delete an Address by AddressId",
+            summary = "Delete an Address by AddressId , but only if the user is an administrator.",
             description = "This API deletes an Address by the provided AddressId and returns a ResponseEntity with a success or failure message.",
             responses = {
                     @ApiResponse(responseCode = "200", description = "Address deleted successfully"),
@@ -103,18 +126,25 @@ public class AddressController
             }
     )
     @DeleteMapping("/Delete")
-    public ResponseEntity<?> Delete(@RequestParam String AddressId)
+    public ResponseEntity<?> Delete(HttpServletRequest request, @RequestParam String AddressId)
     {
         try
         {
-            boolean bool = addressService.Delete(AddressId);
-            if(bool)
+            if(loginSuccess.isSessionExist(request))
             {
-                return ResponseEntity.ok().body(responseDTO.Response("success","데이터베이스에 주소 데이터 삭제 성공"));
+                boolean bool = addressService.Delete(AddressId);
+                if(bool)
+                {
+                    return ResponseEntity.ok().body(responseDTO.Response("success","데이터베이스에 주소 데이터 삭제 성공"));
+                }
+                else
+                {
+                    throw new RuntimeException("예기치 못한 오류로 런타임 오류 발생!!");
+                }
             }
             else
             {
-                throw new RuntimeException("예기치 못한 오류로 런타임 오류 발생!!");
+                throw new SessionNotFoundExpection("관리자 로그인 필요!! 권한이 없어서 데이터 삭제 불가");
             }
         } 
         catch (Exception e) 
@@ -125,7 +155,7 @@ public class AddressController
 
 
     @Operation(
-            summary = "Returns a list of AddressDTO objects",
+            summary = "Returns a list of AddressDTO objects , but only if the user is an administrator.",
             description = "This API retrieves a list of AddressDTO objects from the database.",
             responses = {
                     @ApiResponse(responseCode = "200", description = "Address List Found successfully", content = @Content(mediaType = "application/json", schema = @Schema(implementation = AddressDTO.class))),
@@ -133,22 +163,30 @@ public class AddressController
             }
     )
     @GetMapping
-    public ResponseEntity<?> FindAll()
+    public ResponseEntity<?> FindAll(HttpServletRequest request)
     {
         try
         {
-           List<AddressEntity> addressEntityList = addressService.findALL();
-            if(addressEntityList.isEmpty())
+            if(loginSuccess.isSessionExist(request))
             {
-                throw new FindFailedException("데이터베이스를 조회하긴 했으나, 비어있습니다.");
-            }
-            else {
-                List<AddressDTO> addressDTOS = new ArrayList<>();
-                for (AddressEntity addressEntity : addressEntityList) {
-                    addressDTOS.add(ConvertToDTO(addressEntity));
+                List<AddressEntity> addressEntityList = addressService.findALL();
+                if(addressEntityList.isEmpty())
+                {
+                    throw new FindFailedException("데이터베이스를 조회하긴 했으나, 비어있습니다.");
                 }
-                return ResponseEntity.ok().body(responseDTO.Response("success","데이터베이스에 주소 데이터 조회 성공", addressDTOS));
+                else {
+                    List<AddressDTO> addressDTOS = new ArrayList<>();
+                    for (AddressEntity addressEntity : addressEntityList) {
+                        addressDTOS.add(ConvertToDTO(addressEntity));
+                    }
+                    return ResponseEntity.ok().body(responseDTO.Response("success","데이터베이스에 주소 데이터 조회 성공", addressDTOS));
+                }
             }
+            else
+            {
+                throw new SessionNotFoundExpection("관리자 로그인X, 권한이 없어서 예약 정보 조회 불가");
+            }
+
         }
         catch (Exception e)
         {
@@ -158,7 +196,7 @@ public class AddressController
 
 
     @Operation(
-            summary = "Returns AddressDTO object for a given Address Id",
+            summary = "Returns AddressDTO object for a given Address Id , but only if the user is an administrator.",
             description = "This API retrieves an Address based on the provided Address Id and returns the corresponding AddressDTO object.",
             responses = {
                     @ApiResponse(responseCode = "200", description = "Address found successfully", content = @Content(mediaType = "application/json", schema = @Schema(implementation = AddressDTO.class))),
@@ -166,19 +204,26 @@ public class AddressController
             }
     )
     @PostMapping("/Find")
-    public ResponseEntity<?> FindById(@RequestParam String AddressId)
+    public ResponseEntity<?> FindById(HttpServletRequest request, @RequestParam String AddressId)
     {
         try
         {
-            AddressEntity addressEntity = addressService.findByAddressId(AddressId);
-            if(addressEntity != null)
+            if(loginSuccess.isSessionExist(request))
             {
-                AddressDTO addressDTO = ConvertToDTO(addressEntity);
-                return ResponseEntity.ok().body(responseDTO.Response("success", "데이터 조회 성공", Collections.singletonList(addressDTO)));
+                AddressEntity addressEntity = addressService.findByAddressId(AddressId);
+                if(addressEntity != null)
+                {
+                    AddressDTO addressDTO = ConvertToDTO(addressEntity);
+                    return ResponseEntity.ok().body(responseDTO.Response("success", "데이터 조회 성공", Collections.singletonList(addressDTO)));
+                }
+                else
+                {
+                    throw new RuntimeException("예기치 못한 오류로 런타임 오류 발생!!");
+                }
             }
             else
             {
-                throw new RuntimeException("예기치 못한 오류로 런타임 오류 발생!!");
+                throw new SessionNotFoundExpection("관리자 로그인 X, 예약정보 조회 불가");
             }
         }
         catch (Exception e)
